pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pynika"
version = "0.1.0"
description = "Calibration tool for small-angle and wide-angle pinhole X-ray scattering instruments"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Ivan Ilavsky", email = "ilavsky@aps.anl.gov" },
]
keywords = ["SAXS", "WAXS", "X-ray scattering", "calibration", "synchrotron"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "h5py>=3.8",
]

[project.optional-dependencies]
gui = [
    "PyQt6>=6.4",
    "pyqtgraph>=0.13",
]
epics = [
    "pyepics>=3.5",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]
all = [
    "pynika[gui,epics,dev]",
]

[project.scripts]
pynika = "pynika.cli:main"

[project.urls]
Homepage = "https://github.com/ilavsky/pynika"
Documentation = "https://github.com/ilavsky/pynika/tree/main/doc"
"Bug Tracker" = "https://github.com/ilavsky/pynika/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["pynika*"]

[tool.setuptools.package-data]
pynika = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
